Move column viewset to its own file.
